third_party/WebKit/LayoutTests/platform/linux/virtual/outofblink-cors-http/http/tests/serviceworker/chromium/frame-detached-by-navigation-expected.txt

This is a testharness.js-based test.
FAIL navigator.serviceWorker should return null when frame is detached assert_equals: expected null but got object "[object ServiceWorkerContainer]"
Harness: the test ran to completion.
